Squashed 'cmd/service-catalog/go/src/github.com/kubernetes-incubator/service-catalog/' changes from d969acde90..b69b4a6c80
b69b4a6c80 origin build: modify hard coded path
527fac4d02 origin build: add origin tooling
545ffdbde0 chart changes for v0.1.5 release (#1709)
4d9be8f98e Use userInfo for Originating-Identity so extras is correct. (#1702)
f358b9977d Call destroy function on each storage interface (#1705)
36b5de93c2 refactor binding reconciliation functions (#1687)
5699360b57 Change binding_retrievable to bindingRetrievable
0d8bcfe6fa thread through stopCh to DestroyFunc (#1671)
1c45aef3b1 Migrate from glide to dep for dependency management (#1670)
1cf0dd99c4 Add svcat to Makefile (#1683)
45b1013df4 make verify validates that versioned APIs contain json tags for fields, addresses #1303 (#1480)
0ee83982d0 Build the integration test binary before running any tests (#1666)
0fe0aa7828 Update design.md (#1674)
1280d24b90 controller requires permission to update secrets (#1663)
129d98e6f6 Contribute svcat (#1664)
ff9739b1c4 Update dependencies to kubernetes-1.9.1 (#1633)
9c360198dd chart changes for v0.1.4 release (#1669)
93319f6853 move apiserver generation to script and verify (#1662)
385f0daf1c refactor service instance provision/update/poll reconciliation (#1648)
e015212a01 run each integration test individually (#1661)
412e2427c3 Tell people whether we're checking external hrefs (#1659)
ae053611b2 retry failed unbind requests (#1653)
7eae8458d4 doc for setting up Service Catalog with Prometheus metrics (#1654)
0720cf9595 minor README copy edit (#1656)
8bd347d313 run some integration subtests in parallel (#1637)
b83800c12a Use $ and console to indicate multi-command blocks
789c4b27e4 Use dynamic reaction to fix data race (#1650)
f1be763be9 only check external hrefs on master (#1652)
65c6d20e9d Controller-manager crash loops if API server is not available on startup (#1376) (#1591)
9225c920f9 embedded etcd is the way of the future for our tests (#1651)
605c9520f0 Fix required fields in OpenAPI schema (#1602)
899ca21780 Revert "Switch to wget for integration apiserver checks (#1384)" (#1585)
2f496ee63f Update code-of-conduct.md (#1635)
c1c69cf27b Build the e2e binary in CI (#1647)
4e2dcef04e Wait for successful API discovery (#1646)
5ae6d99b34 Bump copyright date in generated code (#1645)
8be5b05a5c Serve OpenAPI spec only when --serve-openapi-spec switch is enabled (#1612)
19fb30e212 silence go-restful logging output (#1622)
fdbabf036e Add walkthrough link back (#1620)
7c73e9afc9 Add link to main k8s docs on service-catalog (#1627)
f59adc9773 Overhauling the design document (#1619)
cd7b63353e Updating the install documentation (#1616)
f6e5441dba fix compilation error from updated util.WaitForBindingCondition() (#1629) (#1631)
54e57af23a Provide OSB Client proxy to instrument with metrics (#1615)
026b86f803 Disable test added in 1611 that contains data race (#1626)
cb735a6842 Add integration tests for ServiceInstances (#1611)
67dbabbb96 Cleaning up the docs README (#1618)
6bddc0744f remove email from docker login during Travis deploy (#1614)
a604bc398b Use ConfigMaps for leader election (#1599)
c6f193ac23 Add controller integration tests for ServiceInstance create and update (#1578)
26cf23b401 Rename OWNERS assignees: to approvers: (#1508)
1163edc6a5 expose Prometheus metrics from Controller (#677) (#1608)
2cd6554121 Clean up docs/ folder (#1609)
1d7e96d32e Adding Service Binding Create Integration Tests (#1580)
6a4c469197 Make the maximum polling backoff configurable (#1607)
31bbf5539d Rename the imported package to avoid name conflict (#1603)
3cdd556839 Add validation for broker spec to SAR admission controller (#1605)
a3408ce11a fix docker volume mount when building with docker under SELinux (#1500) (#1534)
307e747a5d Remove unneeded vendors of vendors (#1596)
770fc74f3f Make ups-instance.yaml in walkthrough to demonstrate good practices (#1592)
9112ba1953 Add links to docs/README (#1589)
8902648c03 Add additional service to ups-broker to fix e2e (#1583)
0bcbc7d72c move instance update logic out of reconcileServiceInstanceDelete (#1584)
7ef5a3ec94 Do not send Parameters field when there are no parameters from sourced secret (#1559)
4c51b25cd2 Remove unneeded code that sets reason for provision/update call failure (#1561)
b122cb92c2 fix bind injection failure not being persisted in API server (#1546)
66421d590d Clear out current operation when starting reconciliation of a delete (#1564)
8cca70aa85 Send an empty object for Parameters when deleting all parameters of a ServiceInstance (#1555)
270426c5fd Add controllerTest type as a helper for running controller integration tests. (#1577)
e26c2d7bae Ignore .vscode folder in project root (#1579)
REVERT: d969acde90 Add additional service to ups-broker to fix e2e (#1583)
REVERT: 1bcd53b684 origin build: add origin tooling

git-subtree-dir: cmd/service-catalog/go/src/github.com/kubernetes-incubator/service-catalog
git-subtree-split: b69b4a6c8003f25d040e3087c7b1b16d1854a9e9

+++ test/integration/etcd_test.go
@@ -61,8 +61,18 @@ func startEtcd() error {
 }
 
 func stopEtcd() {
+	if etcdContext.etcd == nil {
+		return
+	}
 	etcdContext.etcd.Server.Stop()
 	os.RemoveAll(etcdContext.dir)
+
+	select {
+	case <-etcdContext.etcd.Server.StopNotify():
+		glog.Info("server is stopped!")
+	case <-time.After(60 * time.Second):
+		glog.Error("server took too long to stop!")
+	}
 }
 
 func TestMain(m *testing.M) {
